snowflake_bq_unified_pipeline/requirements.txt
# Core dependencies
functions-framework==3.5.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
google-cloud-monitoring==2.16.0
google-cloud-error-reporting==1.9.3
snowflake-connector-python==3.5.0

# Data processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# Configuration and utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# HTTP and async
requests==2.31.0
urllib3==2.1.0

# Performance
cachetools==5.3.2